requirements.txt
psycopg2>=2.9.3
uWSGI>=2.0.19.1,<2.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5

